test(select): migrate popover-size
